cfn/tagmemo/auth/src/app/requirements.txt
python-decouple
